text-to-video-app/requirements.txt
Flask==2.3.3
moviepy==1.0.3
gtts==2.4.0
pydub==0.25.1
werkzeug==2.3.7
requests==2.31.0
pillow==10.0.1
